# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Agri-PoD Inventory & Distribution System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /farmers:
    get:
      operationId: listFarmers
      tags: [farmers]
      summary: List farmers
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: search
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, approved, rejected]
        - name: districtId
          in: query
          schema:
            type: integer
        - name: valueChainId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Farmers list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedFarmers"
    post:
      operationId: createFarmer
      tags: [farmers]
      summary: Register farmer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFarmerBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Farmer"
  /farmers/{id}:
    get:
      operationId: getFarmer
      tags: [farmers]
      summary: Get farmer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Farmer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Farmer"
    put:
      operationId: updateFarmer
      tags: [farmers]
      summary: Update farmer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateFarmerBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Farmer"
  /farmers/{id}/approve:
    post:
      operationId: approveFarmer
      tags: [farmers]
      summary: Approve farmer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Farmer"
  /farmers/{id}/reject:
    post:
      operationId: rejectFarmer
      tags: [farmers]
      summary: Reject farmer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RejectFarmerBody"
      responses:
        "200":
          description: Rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Farmer"
...
  /inventory/input-items:
    get:
      operationId: listInputItems
      tags: [inventory]
      summary: List input items catalogue
      responses:
        "200":
          description: Input items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InputItem"
    post:
      operationId: createInputItem
      tags: [inventory]
      summary: Create input item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInputItemBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InputItem"
  /inventory/stock-balance:
    get:
      operationId: getStockBalance
      tags: [inventory]
      summary: Get stock balance across warehouses
      parameters:
        - name: warehouseId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Stock balance
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StockBalance"
  /inventory/receive-stock:
    post:
      operationId: receiveStock
      tags: [inventory]
      summary: Receive stock into warehouse
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReceiveStockBody"
      responses:
        "201":
          description: Stock received
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StockTransaction"
  /inventory/transfer-stock:
    post:
      operationId: transferStock
      tags: [inventory]
      summary: Transfer stock between warehouses
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransferStockBody"
      responses:
        "201":
          description: Stock transferred
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StockTransaction"
...
  /procurement:
    get:
      operationId: listProcurementOrders
      tags: [procurement]
      summary: List procurement orders
      responses:
        "200":
          description: Procurement orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProcurementOrder"
    post:
      operationId: createProcurementOrder
      tags: [procurement]
      summary: Create procurement order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProcurementOrderBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProcurementOrder"
  /procurement/{id}:
    get:
      operationId: getProcurementOrder
      tags: [procurement]
      summary: Get procurement order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Procurement order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProcurementOrder"

  # CAMPAIGNS
...
components:
  schemas:
...
    Farmer:
      type: object
      properties:
        id:
          type: integer
        farmerCode:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        gender:
          type: string
          enum: [Male, Female, Other]
        phone:
          type: string
          nullable: true
        nationalId:
          type: string
          nullable: true
        districtId:
          type: integer
        districtName:
          type: string
          nullable: true
        chiefdomId:
          type: integer
          nullable: true
        chiefdomName:
          type: string
          nullable: true
        sectionId:
          type: integer
          nullable: true
        communityId:
          type: integer
          nullable: true
        communityName:
          type: string
          nullable: true
        valueChainId:
          type: integer
        valueChainName:
          type: string
          nullable: true
        farmSize:
          type: number
          nullable: true
        gpsLatitude:
          type: number
          nullable: true
        gpsLongitude:
          type: number
          nullable: true
        photoUrl:
          type: string
          nullable: true
        status:
          type: string
          enum: [pending, approved, rejected]
        barcodeToken:
          type: string
          nullable: true
        ageGroup:
          type: string
          nullable: true
        farmerGroup:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
          nullable: true
      required: [id, farmerCode, firstName, lastName, gender, districtId, valueChainId, status, createdAt]
    CreateFarmerBody:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        gender:
          type: string
          enum: [Male, Female, Other]
        phone:
          type: string
          nullable: true
        nationalId:
          type: string
          nullable: true
        districtId:
          type: integer
        chiefdomId:
          type: integer
          nullable: true
        sectionId:
          type: integer
          nullable: true
        communityId:
          type: integer
          nullable: true
        valueChainId:
          type: integer
        farmSize:
          type: number
          nullable: true
        gpsLatitude:
          type: number
          nullable: true
        gpsLongitude:
          type: number
          nullable: true
        ageGroup:
          type: string
          nullable: true
        farmerGroup:
          type: string
          nullable: true
      required: [firstName, lastName, gender, districtId, valueChainId]
    UpdateFarmerBody:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        gender:
          type: string
          enum: [Male, Female, Other]
        phone:
          type: string
          nullable: true
        districtId:
          type: integer
        valueChainId:
          type: integer
        farmSize:
          type: number
          nullable: true
    RejectFarmerBody:
      type: object
      properties:
        reason:
          type: string
      required: [reason]
    PaginatedFarmers:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Farmer"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required: [data, total, page, limit, totalPages]
...
    InputItem:
      type: object
      properties:
        id:
          type: integer
        code:
          type: string
        name:
          type: string
        category:
          type: string
        valueChainId:
          type: integer
          nullable: true
        valueChainName:
          type: string
          nullable: true
        unitOfMeasure:
          type: string
        description:
          type: string
          nullable: true
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [id, code, name, category, unitOfMeasure, isActive, createdAt]
    CreateInputItemBody:
      type: object
      properties:
        name:
          type: string
        category:
          type: string
        valueChainId:
          type: integer
          nullable: true
        unitOfMeasure:
          type: string
        description:
          type: string
          nullable: true
      required: [name, category, unitOfMeasure]
    StockBalance:
      type: object
      properties:
        warehouseId:
          type: integer
        warehouseName:
          type: string
        inputItemId:
          type: integer
        inputItemName:
          type: string
        inputItemCode:
          type: string
        unitOfMeasure:
          type: string
        batchNumber:
          type: string
          nullable: true
        available:
          type: integer
        reserved:
          type: integer
        loaded:
          type: integer
        delivered:
          type: integer
        returned:
          type: integer
        damaged:
          type: integer
        total:
          type: integer
      required: [warehouseId, warehouseName, inputItemId, inputItemName, inputItemCode, unitOfMeasure, available, reserved, loaded, delivered, returned, damaged, total]
    ReceiveStockBody:
      type: object
      properties:
        warehouseId:
          type: integer
        inputItemId:
          type: integer
        quantity:
          type: integer
        batchNumber:
          type: string
          nullable: true
        expiryDate:
          type: string
          format: date
          nullable: true
        supplierId:
          type: integer
          nullable: true
        procurementOrderId:
          type: integer
          nullable: true
        notes:
          type: string
          nullable: true
      required: [warehouseId, inputItemId, quantity]
    TransferStockBody:
      type: object
      properties:
        fromWarehouseId:
          type: integer
        toWarehouseId:
          type: integer
        inputItemId:
          type: integer
        quantity:
          type: integer
        notes:
          type: string
          nullable: true
      required: [fromWarehouseId, toWarehouseId, inputItemId, quantity]
    StockTransaction:
      type: object
      properties:
        id:
          type: integer
        transactionType:
          type: string
        warehouseId:
          type: integer
        warehouseName:
          type: string
          nullable: true
        inputItemId:
          type: integer
        inputItemName:
          type: string
          nullable: true
        quantity:
          type: integer
        batchNumber:
          type: string
          nullable: true
        status:
          type: string
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        createdBy:
          type: string
          nullable: true
      required: [id, transactionType, warehouseId, inputItemId, quantity, status, createdAt]
...
    ProcurementOrder:
      type: object
      properties:
        id:
          type: integer
        orderCode:
          type: string
        supplierName:
          type: string
        inputItemId:
          type: integer
        inputItemName:
          type: string
          nullable: true
        quantity:
          type: integer
        unitPrice:
          type: number
          nullable: true
        totalValue:
          type: number
          nullable: true
        fundingSource:
          type: string
          nullable: true
        status:
          type: string
        warehouseId:
          type: integer
          nullable: true
        warehouseName:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        orderDate:
          type: string
          format: date
        createdAt:
          type: string
          format: date-time
      required: [id, orderCode, supplierName, inputItemId, quantity, status, orderDate, createdAt]
    CreateProcurementOrderBody:
      type: object
      properties:
        supplierName:
          type: string
        inputItemId:
          type: integer
        quantity:
          type: integer
        unitPrice:
          type: number
          nullable: true
        fundingSource:
          type: string
          nullable: true
        warehouseId:
          type: integer
          nullable: true
        orderDate:
          type: string
          format: date
        notes:
          type: string
          nullable: true
      required: [supplierName, inputItemId, quantity, orderDate]

    # CAMPAIGNS
...